[12174] Finalize first part of Vehicle Implementation
Special thanks to kid10 for lasting endurance to get this finally ready to go

Also thanks to the very many people to contributed over the years to make this possible.

Just to name a few:
NetSky for first version, Rastikzzz for further development, Wojta and the Valhalla team for further work.
Zergtmn for streamlinging the design, rsa for maintaining a usable version over the time.
Machiavelli from Trinity for ongoing research (especially about flags)

Thanks to crackm for providing dataz, and to the people who actually contributed the data to the developers for implementation

Thanks to the many people who helped with testing to get nasty little bug fixed!

Happy driving with C-Mangos :)

Signed-off-by: Schmoozerd <schmoozerd@cmangos>