[12338] Calendar implementation.
Thank to Spp (trinity dev) for initial research and implementation. Thank to Horn, Warpten and maybe some other for the rework of the original version.

Thank olso to
Schmoozerd for his presence when i doubt for something and for some cleaning :)
Vladimir, Kid10, RSA, xfurry, X-Savior for their advices or sniffs.